requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pillow==11.0.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
